refactor(src/): Update imports for `Wallet.tsx`
