+++ client/src/test/components/common/Input.test.tsx
@@ -1,6 +1,6 @@
 import { Web3ReactProvider } from "@web3-react/core";
 import Input from "components/common/Input";
-import { getLibrary } from "components/Wallet";
+import { getLibrary } from "components/wallet/Wallet";
 import NetworkProvider from "context/network/NetworkProvider";
 import ThemeProvider from "context/theme/ThemeProvider";
 import { mount } from "enzyme";
