+++ client/src/test/components/common/MobileMenu.test.tsx
@@ -8,7 +8,7 @@ import {
 } from "@testing-library/react";
 import { Web3ReactProvider } from "@web3-react/core";
 import MobileMenu from "components/common/MobileMenu";
-import { getLibrary } from "components/Wallet";
+import { getLibrary } from "components/wallet/Wallet";
 import NetworkProvider from "context/network/NetworkProvider";
 import ThemeProvider from "context/theme/ThemeProvider";
 import { mount } from "enzyme";
