+++ client/src/test/components/common/Networks.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render } from "@testing-library/react";
 import { Web3ReactProvider } from "@web3-react/core";
 import NetworkSpeedDial from "components/common/Networks";
-import { getLibrary } from "components/Wallet";
+import { getLibrary } from "components/wallet/Wallet";
 import NetworkProvider from "context/network/NetworkProvider";
 import ThemeProvider from "context/theme/ThemeProvider";
 import { mount } from "enzyme";
