+++ client/src/test/components/create-collection-form/tabs/TabPanel.test.tsx
@@ -1,6 +1,6 @@
 import { Web3ReactProvider } from "@web3-react/core";
 import TabPanel from "components/create-collection-form/tabs/TabPanel";
-import { getLibrary } from "components/Wallet";
+import { getLibrary } from "components/wallet/Wallet";
 import NetworkProvider from "context/network/NetworkProvider";
 import ThemeProvider from "context/theme/ThemeProvider";
 import { mount } from "enzyme";
